setup: Configure RTK stuff
