+++ src/app/page.tsx
@@ -1,10 +1,9 @@
-import { HelloWorld } from '@/src/components/hello-world';
 import type { Metadata } from 'next';
 
 export default function MainPage() {
-  return <HelloWorld />;
+  return <div>'Payment checkout'</div>;
 }
 
 export const metadata: Metadata = {
-  title: 'Hello World',
+  title: 'Payment checkout',
 };
