--- src/components/hello-world.tsx
@@ -1,13 +1,0 @@
-export const HelloWorld = () => {
-  return (
-    <div
-      style={{
-        display: 'flex',
-        justifyContent: 'center',
-        alignItems: 'center',
-      }}
-    >
-      <h3>Hello World!</h3>
-    </div>
-  );
-};
