+++ src/constants/slice-names-constants.ts
@@ -1,0 +1,4 @@
+export const sliceNamesConstants = {
+  paymentCheckoutSlice: 'paymentCheckoutSlice',
+  paymentCheckoutApi: 'paymentCheckoutApi',
+};
