--- src/lib/store.ts
@@ -1,29 +1,0 @@
-import type { Action, ThunkAction } from '@reduxjs/toolkit';
-import { combineSlices, configureStore } from '@reduxjs/toolkit';
-import helloWorldReducer from './features/hello-world-slice';
-import { helloWorldApi } from '@/src/lib/features/hello-world-api';
-
-const rootReducer = combineSlices({
-  helloWorld: helloWorldReducer,
-  [helloWorldApi.reducerPath]: helloWorldApi.reducer,
-});
-
-export type RootState = ReturnType<typeof rootReducer>;
-
-export const makeStore = () => {
-  return configureStore({
-    reducer: rootReducer,
-    middleware: (getDefaultMiddleware) => {
-      return getDefaultMiddleware().concat(helloWorldApi.middleware);
-    },
-  });
-};
-
-export type AppStore = ReturnType<typeof makeStore>;
-export type AppDispatch = AppStore['dispatch'];
-export type AppThunk<ThunkReturnType = void> = ThunkAction<
-  ThunkReturnType,
-  RootState,
-  unknown,
-  Action
->;
