+++ src/redux/store/index.ts
@@ -1,0 +1,24 @@
+import type { Action, ThunkAction } from '@reduxjs/toolkit';
+import { configureStore } from '@reduxjs/toolkit';
+import { paymentCheckoutApi } from '@/src/redux/payment-checkout/payment-checkout-api';
+import { rootReducer } from '@/src/redux/store/root-reducer';
+
+export type RootState = ReturnType<typeof rootReducer>;
+
+export const makeStore = () => {
+  return configureStore({
+    reducer: rootReducer,
+    middleware: getDefaultMiddleware => {
+      return getDefaultMiddleware().concat(paymentCheckoutApi.middleware);
+    },
+  });
+};
+
+export type AppStore = ReturnType<typeof makeStore>;
+export type AppDispatch = AppStore['dispatch'];
+export type AppThunk<ThunkReturnType = void> = ThunkAction<
+  ThunkReturnType,
+  RootState,
+  unknown,
+  Action
+>;
