--- src/types/hello-world.d.ts
@@ -1,3 +1,0 @@
-export interface HelloWorld {
-  value: string;
-}
