+++ src/types/payment-checkout.d.ts
@@ -1,0 +1,3 @@
+export interface PaymentCheckoutState {
+  value: string;
+}
